pyproject.toml
[project]
name = "flaresolverr"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aioquic==1.2.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asgiref==3.10.0",
    "attrs==26.1.0",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.3",
    "blinker==1.9.0",
    "brotli==1.2.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "cryptography==46.0.6",
    "flask==3.1.2",
    "func-timeout==4.3.5",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "kaitaistruct==0.11",
    "ldap3==2.9.1",
    "legacy-cgi==2.6.4",
    "markupsafe==3.0.3",
    "mitmproxy==12.2.1",
    "mitmproxy-linux==0.12.9",
    "mitmproxy-rs==0.12.9",
    "msgpack==1.1.2",
    "outcome==1.3.0.post0",
    "packaging==24.0",
    "prometheus-client==0.23.1",
    "publicsuffix2==2.20191221",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pylsqpack==0.3.23",
    "pyopenssl==25.3.0",
    "pyparsing==3.2.5",
    "pyperclip==1.11.0",
    "pysocks==1.7.1",
    "python-dotenv==1.2.2",
    "requests==2.32.5",
    "ruamel-yaml==0.18.16",
    "selenium==4.41.0",
    "selenium-wire-lw==3.0.2",
    "selenium-wire-undetected-chromedriver-lw==1.0.1",
    "service-identity==24.2.0",
    "setuptools==82.0.1",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.8.3",
    "tornado==6.5.2",
    "trio==0.33.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.15.0",
    "undetected",
    "undetected-chromedriver==3.5.5",
    "urllib3==2.6.3",
    "urwid==3.0.3",
    "uvicorn==0.37.0",
    "wcwidth==0.6.0",
    "webdriver-manager==4.0.2",
    "webob==1.8.9",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "webtest==3.0.7",
    "werkzeug==3.1.7",
    "wsproto==1.2.0",
    "xvfbwrapper==0.2.16",
    "zstandard==0.25.0",
    "fastapi==0.117.1",
    "zstd==1.5.7.3",
]

[tool.uv.sources]
undetected-chromedriver = { git = "https://www.github.com/ultrafunkamsterdam/undetected-chromedriver", rev = "master" }
undetected = { git = "https://github.com/carl0smat3us/undetected" }
